# Fix data races between cq-shutdown and cq-begin-op for a pluckable cq