Also some misc cleanup.