@sreecha 2017-08-03
We saw a tsan error in one of the internal tests Yuchen ran (the race was on GPR_ASSERT(cqd->shutdown_called) line that I removed now). 

Btw, since pluck is very heavily used in sync api, I will run the sync benchmarks before merging this.