@sreecha 2017-08-09
Before this change:
```
Benchmark              Time           CPU Iterations
----------------------------------------------------
BM_Pluck1Core        203 ns        202 ns    3445716 locks/iter:2 atm_cas/iter:0 atm_add/iter:5 nows/iter:0 allocs/iter:2.90215e-07
```

After this change:
```
Benchmark              Time           CPU Iterations
----------------------------------------------------
BM_Pluck1Core        205 ns        205 ns    3416548 locks/iter:2 atm_cas/iter:1 atm_add/iter:4 nows/iter:0 allocs/iter:2.92693e-07
```

One atm_add is converted to atm_cas