@sreecha 2017-08-09
Ran the `cpp_protobuf_async_client_sync_server_streaming_qps_unconstrained_secure` benchmark on my machine with and without this change and there is no visible difference in the QPS

QPS: 153184.7  (without this change)
QPS: 150421.4  (with this change)